Fix PORadioButtonStyle env propagation
