+++ Sources/ProcessOutCoreUI/Sources/DesignSystem/RadioButton/PORadioButtonStyle.swift
@@ -39,31 +39,33 @@ public struct PORadioButtonStyle: ButtonStyle {
     // MARK: - ButtonStyle
 
     public func makeBody(configuration: Configuration) -> some View {
-        let style = currentStyle(isPressed: configuration.isPressed)
-        HStack(alignment: .top, spacing: Constants.spacing) {
-            ZStack {
-                Circle()
-                    .fill(Color(style.knob.backgroundColor))
-                Circle()
-                    .strokeBorder(Color(style.knob.border.color), lineWidth: style.knob.border.width)
-                Circle()
-                    .fill(Color(style.knob.innerCircleColor))
-                    .frame(width: style.knob.innerCircleRadius * 2)
+        ContentView { isSelected, isInvalid, sizeMultipler in
+            let style = currentStyle(isSelected: isSelected, isInvalid: isInvalid, isPressed: configuration.isPressed)
+            HStack(alignment: .top, spacing: Constants.spacing) {
+                ZStack {
+                    Circle()
+                        .fill(Color(style.knob.backgroundColor))
+                    Circle()
+                        .strokeBorder(Color(style.knob.border.color), lineWidth: style.knob.border.width)
+                    Circle()
+                        .fill(Color(style.knob.innerCircleColor))
+                        .frame(width: style.knob.innerCircleRadius * 2)
+                }
+                .frame(width: Constants.knobSize, height: Constants.knobSize)
+                .offset(
+                    y: knobVerticalOffset(typography: style.value.typography, sizeMultipler: sizeMultipler)
+                )
+                configuration
+                    .label
+                    .textStyle(style.value)
+                    .frame(minHeight: Constants.knobSize)
             }
-            .frame(width: Constants.knobSize, height: Constants.knobSize)
-            .offset(
-                y: (knobVerticalOffset(typography: style.value.typography) - Constants.knobSize) / 2
-            )
-            configuration
-                .label
-                .textStyle(style.value)
-                .frame(minHeight: Constants.knobSize)
+            .padding(.vertical, Constants.minVerticalPadding)
+            .frame(maxWidth: .infinity, minHeight: Constants.minHeight, alignment: .leading)
+            .animation(.default, value: isSelected)
+            .contentShape(.rect)
+            .allowsHitTesting(!isSelected)
         }
-        .padding(.vertical, Constants.minVerticalPadding)
-        .frame(maxWidth: .infinity, minHeight: Constants.minHeight, alignment: .leading)
-        .animation(.default, value: isSelected)
-        .contentShape(.rect)
-        .allowsHitTesting(!isSelected)
     }
 
     // MARK: - Private Nested Types
@@ -75,18 +77,9 @@ public struct PORadioButtonStyle: ButtonStyle {
         static let knobSize: CGFloat = 18
     }
 
-    // MARK: - Private Properties
-
-    @Environment(\.isRadioButtonSelected) private var isSelected
-    @Environment(\.isControlInvalid) private var isInvalid
-    @Environment(\.sizeCategory) private var sizeCategory
-
-    @POBackport.ScaledMetric
-    private var contentSizeMultipler: CGFloat = 1
-
     // MARK: - Private Methods
 
-    private func currentStyle(isPressed: Bool) -> PORadioButtonStateStyle {
+    private func currentStyle(isSelected: Bool, isInvalid: Bool, isPressed: Bool) -> PORadioButtonStateStyle {
         if isSelected {
             return selected
         }
@@ -99,60 +92,29 @@ public struct PORadioButtonStyle: ButtonStyle {
         return normal
     }
 
-    private func knobVerticalOffset(typography: POTypography) -> CGFloat {
-        if typography.textStyle != nil {
-            return typography.lineHeight * _contentSizeMultipler.value(relativeTo: typography.textStyle)
-        }
-        return typography.lineHeight
+    private func knobVerticalOffset(
+        typography: POTypography, sizeMultipler: POBackport<Any>.ScaledMetric<CGFloat>
+    ) -> CGFloat {
+        let lineHeight = typography.lineHeight * sizeMultipler.value(relativeTo: typography.textStyle)
+        return (lineHeight - Constants.knobSize) / 2
     }
 }
 
-extension ButtonStyle where Self == PORadioButtonStyle {
+// Environments are not propagated directly to ButtonStyle in any iOS before 14.
+private struct ContentView<Content: View>: View {
 
-    @_spi(PO) public static var radio: PORadioButtonStyle {
-        PORadioButtonStyle(
-            normal: .init(
-                knob: .init(
-                    backgroundColor: .clear,
-                    border: .regular(color: UIColor(resource: .Border.default)),
-                    innerCircleColor: .clear,
-                    innerCircleRadius: 0
-                ),
-                value: valueStyle
-            ),
-            selected: .init(
-                knob: .init(
-                    backgroundColor: .clear,
-                    border: .regular(color: UIColor(resource: .Action.Primary.default)),
-                    innerCircleColor: UIColor(resource: .Action.Primary.default),
-                    innerCircleRadius: 4
-                ),
-                value: valueStyle
-            ),
-            highlighted: .init(
-                knob: .init(
-                    backgroundColor: .clear,
-                    border: .regular(color: UIColor(resource: .Text.muted)),
-                    innerCircleColor: .clear,
-                    innerCircleRadius: 0
-                ),
-                value: valueStyle
-            ),
-            error: .init(
-                knob: .init(
-                    backgroundColor: .clear,
-                    border: .regular(color: UIColor(resource: .Text.error)),
-                    innerCircleColor: .clear,
-                    innerCircleRadius: 0
-                ),
-                value: valueStyle
-            )
-        )
+    @ViewBuilder
+    let content: (_ isSelected: Bool, _ isInvalid: Bool, POBackport<Any>.ScaledMetric<CGFloat>) -> Content
+
+    var body: some View {
+        content(isSelected, isInvalid, _contentSizeMultipler)
     }
 
     // MARK: - Private Properties
 
-    private static var valueStyle: POTextStyle {
-        POTextStyle(color: UIColor(resource: .Text.primary), typography: .Fixed.label)
-    }
+    @Environment(\.isRadioButtonSelected) private var isSelected
+    @Environment(\.isControlInvalid) private var isInvalid
+
+    @POBackport.ScaledMetric
+    private var contentSizeMultipler: CGFloat = 1
 }
